change wysiwyg strategy namespaces change installation and configuration steps in docs
