+++ doc/01.2-webpack-entry.md
@@ -15,9 +15,12 @@
 
 // Shop config
     .addEntry('bitbag-cms-shop', 'vendor/bitbag/cms-plugin/src/Resources/assets/shop/entry.js')
-
-// Admin config
+    
+// Admin config - add one of the following entries (CKeditor or Trix)
+    // CKeditor
     .addEntry('bitbag-cms-admin', 'vendor/bitbag/cms-plugin/src/Resources/assets/admin/entry.js')
+    // Trix
+    .addEntry('bitbag-cms-admin', 'vendor/bitbag/cms-plugin/src/Resources/assets/admin/trix-entry.js')
 ```
 
 2. Add encore functions to your templates
