+++ doc/01.3-import-entry.md
@@ -16,6 +16,9 @@
 // ./assets/shop/entry.js
 import '../../vendor/bitbag/cms-plugin/src/Resources/assets/shop/entry.js';
 
-// ./assets/admin/entry.js
+// ./assets/admin/entry.js - choose one of the following entries (CKeditor or Trix)
+//CkEditor
 import '../../vendor/bitbag/cms-plugin/src/Resources/assets/admin/entry.js';
+// Trix
+import '../../vendor/bitbag/cms-plugin/src/Resources/assets/admin/trix-entry.js';
 ```
