+++ src/Form/Strategy/Wysiwyg/AbstractWysiwygStrategy.php
@@ -8,9 +8,8 @@
 
 declare(strict_types=1);
 
-namespace BitBag\SyliusCmsPlugin\Form\Type\Wysiwyg\Strategy;
+namespace BitBag\SyliusCmsPlugin\Form\Strategy\Wysiwyg;
 
-use BitBag\SyliusCmsPlugin\Form\Type\Wysiwyg\WysiwygStrategyInterface;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
