+++ src/Form/Strategy/Wysiwyg/WysiwygStrategyInterface.php
@@ -8,20 +8,19 @@
 
 declare(strict_types=1);
 
-namespace BitBag\SyliusCmsPlugin\Form\Type\Wysiwyg\Strategy;
+namespace BitBag\SyliusCmsPlugin\Form\Strategy\Wysiwyg;
 
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormView;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class TrixStrategy extends AbstractWysiwygStrategy
+interface WysiwygStrategyInterface
 {
-    public function buildView(FormView $view, FormInterface $form, array $options): void
-    {
-        $view->vars['block_prefix'] = 'bitbag_sylius_cms_plugin_trix_strategy';
-    }
+    public function configureOptions(OptionsResolver $resolver): void;
 
-    public function getBlockPrefix(): string
-    {
-        return 'bitbag_sylius_cms_plugin_trix_strategy';
-    }
+    public function buildView(FormView $view, FormInterface $form, array $options): void;
+
+    public function getParent(): string;
+
+    public function getBlockPrefix(): string;
 }
