+++ src/Resolver/WysiwygStrategyResolver.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace BitBag\SyliusCmsPlugin\Resolver;
 
-use BitBag\SyliusCmsPlugin\Form\Type\Wysiwyg\WysiwygStrategyInterface;
+use BitBag\SyliusCmsPlugin\Form\Strategy\Wysiwyg\WysiwygStrategyInterface;
 
-class WysiwygStrategyResolver implements WysiwygStrategyResolverInterface
+final class WysiwygStrategyResolver implements WysiwygStrategyResolverInterface
 {
     public function __construct(
         private array $strategies,
