+++ src/Resolver/WysiwygStrategyResolverInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace BitBag\SyliusCmsPlugin\Resolver;
 
-use BitBag\SyliusCmsPlugin\Form\Type\Wysiwyg\WysiwygStrategyInterface;
+use BitBag\SyliusCmsPlugin\Form\Strategy\Wysiwyg\WysiwygStrategyInterface;
 
 interface WysiwygStrategyResolverInterface
 {
