+++ src/Resources/assets/admin/trix-entry.js
@@ -1,0 +1,2 @@
+import './js/bitbag/bitbag-trix-editor';
+import './scss/main.scss';
