+++ src/Resources/config/services/form.xml
@@ -82,6 +82,16 @@
             </call>
             <tag name="form.type" />
         </service>
+
+        <service id="bitbag_sylius_cms_plugin.form.wysiwyg_strategy.ckeditor"
+                 class="BitBag\SyliusCmsPlugin\Form\Strategy\Wysiwyg\CkeditorStrategy"
+        >
+            <argument type="service" id="router.default" />
+        </service>
+
+        <service id="bitbag_sylius_cms_plugin.form.wysiwyg_strategy.trix"
+                 class="BitBag\SyliusCmsPlugin\Form\Strategy\Wysiwyg\TrixStrategy"
+        />
     </services>
 
 </container>
