+++ tests/Application/config/packages/bitbag_sylius_cms_plugin.yaml
@@ -4,6 +4,11 @@ imports:
 parameters:
     fixtures_dir: "%kernel.project_dir%/Resources/fixtures"
 
+# Switch WYSIWYG editor to Trix if you want to use it.
+# (Default is CKEditor)
+#bit_bag_sylius_cms:
+#    wysiwyg_editor: trix
+
 sylius_fixtures:
     suites:
         default:
