+++ tests/Application/config/packages/twig.yaml
@@ -5,7 +5,6 @@ twig:
     form_themes:
         - '@FOSCKEditor/Form/ckeditor_widget.html.twig'
         - '@BitBagSyliusCmsPlugin/Form/ckeditor_widget.html.twig'
-        - "@BitBagSyliusCmsPlugin/Widget/_trixWidget.html.twig"
 
 services:
     _defaults:
